move ComposeResult type into common package to allow its reusing
